+++ src/main/resources/scripts/popup-script.js
@@ -1,1 +1,9 @@
 runOnPopup();
+
+$(function() {
+    $("#whichTechnologiesToSee").select2()
+
+    $("#selectedTechnologiesClean").on('click', function (event) {
+        $("#whichTechnologiesToSee").val(null).trigger('change')
+    })
+});