+++ src/main/scala/com/lcampos/model/StorageKeys.scala
@@ -1,5 +1,0 @@
-package com.lcampos.model
-
-object StorageKeys {
-  val isExtensionActive = "recruiter-star-is-active"
-}